requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
db-dtypes==1.1.1
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.12.0
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
mysql-connector-python==8.2.0
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
pandas-gbq==0.19.2
proto-plus==1.22.3
protobuf==4.21.12
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.2
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
